requirements.txt
asn1tools~=0.165.0
requests~=2.28.2
paho-mqtt~=1.6.1